requirements.txt.txt
threading
socket
mysql.connector==2.2.9
sys